+++ src/app/app.js
@@ -16,7 +16,7 @@
 import CsvGrid from '../grid/grid.js';
 import { cleanCsvText } from '../grid/core.js';
 
-const VERSION = '3.0.3';
+const VERSION = '3.0.4';
 
 const $ = id => document.getElementById(id);